어노테이터 대시보드 테이블 컴포넌트 분리
