+++ src/pages/AnnotatorPage/AnnotatorDashboard.css
@@ -214,99 +214,4 @@
 
 .view-details-btn:hover {
   background-color: #3a68cc;
-}
-
-/* Annotation Table */
-.annotation-table-container {
-  width: 1109px;
-  background: white;
-  border-radius: 14px;
-  border: 0.3px solid #B9B9B9;
-  overflow: hidden;
-}
-
-.annotation-table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.annotation-table thead {
-  background: #FCFDFD;
-  border-bottom: 0.6px solid #D5D5D5;
-  height: 49px;
-}
-
-.annotation-table th {
-  padding: 15px;
-  text-align: left;
-  color: #202224;
-  font-size: 14px;
-  font-family: 'Nunito Sans', sans-serif;
-  font-weight: 800;
-  opacity: 0.9;
-}
-
-.annotation-table tbody tr {
-  height: 76px;
-  border-bottom: 0.4px solid #979797;
-  cursor: pointer;
-}
-
-.annotation-table tbody tr:hover {
-  background: #f9f9f9;
-}
-
-.annotation-table td {
-  padding: 15px;
-  color: #202224;
-  font-size: 14px;
-  font-family: 'Nunito Sans', sans-serif;
-  font-weight: 600;
-  opacity: 0.9;
-}
-
-.status-tag {
-  display: inline-flex;
-  align-items: center;
-  padding: 4px 10px;
-  border-radius: 12px;
-  font-size: 12px;
-  font-weight: 600;
-}
-
-.checkbox-col {
-  width: 40px;
-}
-
-.annotation-table input[type="checkbox"] {
-  width: 15.55px;
-  height: 16px;
-  border-radius: 3px;
-  border: 1.2px solid #B3B3B3;
-  cursor: pointer;
-}
-
-.delete-btn {
-  width: 138px;
-  height: 37px;
-  background: #E2EAF8;
-  border-radius: 12px;
-  border: none;
-  cursor: pointer;
-  color: #EF3826;
-  font-size: 14px;
-  font-family: 'Nunito Sans', sans-serif;
-  font-weight: 700;
-  opacity: 0.7;
-}
-
-.delete-btn:hover {
-  opacity: 1;
-}
-
-.no-data-message {
-  text-align: center;
-  padding: 30px 0;
-  color: #718EBF;
-  font-style: italic;
 }
