+++ src/pages/AnnotatorPage/AnnotatorDashboard.jsx
@@ -12,9 +12,10 @@ import {
   STATUS_FILTERS, 
   CONFIDENCE_SCORE_FILTERS 
 } from '../../constants/annotatorDashboardConstants';
-import { formatConfidenceScore, getStatusStyles, extractNumericId } from '../../utils/annotatorDashboardUtils';
+import { extractNumericId } from '../../utils/annotatorDashboardUtils';
 import DashboardSidebar from '../../components/Annotator/Sidebar/DashboardSidebar';
 import DashboardHeader from '../../components/Annotator/Header/DashboardHeader';
+import AnnotationTable from '../../components/Annotator/Table/AnnotationTable';
 import './AnnotatorDashboard.css';
 
 const AnnotatorDashboard = () => {
@@ -38,11 +39,6 @@ const AnnotatorDashboard = () => {
     } else {
       console.error(`Invalid ID format: ${id}`);
     }
-  };
-  
-  const handleDeleteClick = (id, event) => {
-    event.stopPropagation();
-    handleDelete(id);
   };
   
   if (isLoading) {
@@ -77,7 +73,7 @@ const AnnotatorDashboard = () => {
             <h2 className="section-title">Current Task</h2>
             <div className="error-container">
               <p className="error-message">{error}</p>
-              <button className="retry-btn" onClick={refreshData}>다시 시도</button>
+              <button className="retry-btn" onClick={refreshData}>Retry</button>
             </div>
           </div>
         </div>
@@ -128,15 +124,13 @@ const AnnotatorDashboard = () => {
                 className="filter-input"
                 onClick={() => {
                   // 추후 드롭다운 UI 구현
-                  const newValue = window.prompt('결함 유형을 선택하세요: all, scratch, dent, discoloration, contamination', filters[FILTER_TYPES.DEFECT_TYPE]);
+                  const newValue = window.prompt('Select defect type: all, scratch, dent, discoloration, contamination', filters[FILTER_TYPES.DEFECT_TYPE]);
                   if (newValue) {
                     handleFilterChange(FILTER_TYPES.DEFECT_TYPE, newValue);
                   }
                 }}
               >
-                <span>
-                  {DEFECT_TYPE_FILTERS.find(f => f.id === filters[FILTER_TYPES.DEFECT_TYPE])?.label || '모든 유형'}
-                </span>
+                <span>Defect Type</span>
                 <FaChevronDown />
               </div>
             </div>
@@ -146,15 +140,13 @@ const AnnotatorDashboard = () => {
                 className="filter-input"
                 onClick={() => {
                   // 추후 드롭다운 UI 구현
-                  const newValue = window.prompt('상태를 선택하세요: all, completed, pending', filters[FILTER_TYPES.STATUS]);
+                  const newValue = window.prompt('Select status: all, completed, pending', filters[FILTER_TYPES.STATUS]);
                   if (newValue) {
                     handleFilterChange(FILTER_TYPES.STATUS, newValue);
                   }
                 }}
               >
-                <span>
-                  {STATUS_FILTERS.find(f => f.id === filters[FILTER_TYPES.STATUS])?.label || '모든 상태'}
-                </span>
+                <span>Status</span>
                 <FaChevronDown />
               </div>
             </div>
@@ -164,15 +156,13 @@ const AnnotatorDashboard = () => {
                 className="filter-input"
                 onClick={() => {
                   // 추후 드롭다운 UI 구현
-                  const newValue = window.prompt('신뢰도 점수를 선택하세요: all, high, medium, low', filters[FILTER_TYPES.CONFIDENCE_SCORE]);
+                  const newValue = window.prompt('Select confidence score: all, high, medium, low', filters[FILTER_TYPES.CONFIDENCE_SCORE]);
                   if (newValue) {
                     handleFilterChange(FILTER_TYPES.CONFIDENCE_SCORE, newValue);
                   }
                 }}
               >
-                <span>
-                  {CONFIDENCE_SCORE_FILTERS.find(f => f.id === filters[FILTER_TYPES.CONFIDENCE_SCORE])?.label || '모든 점수'}
-                </span>
+                <span>Confidence Score</span>
                 <FaChevronDown />
               </div>
             </div>
@@ -181,63 +171,11 @@ const AnnotatorDashboard = () => {
           </div>
           
           {/* Annotation Table */}
-          <div className="annotation-table-container">
-            <table className="annotation-table">
-              <thead>
-                <tr>
-                  <th className="checkbox-col">
-                    <input type="checkbox" />
-                  </th>
-                  <th>Camera</th>
-                  <th>DATA ID</th>
-                  <th>CONFIDENCE SCORE(MIN)</th>
-                  <th>COUNT</th>
-                  <th>STATUS</th>
-                  <th>Actions</th>
-                </tr>
-              </thead>
-              <tbody>
-                {annotations.map((annotation) => (
-                  <tr key={annotation.id} onClick={() => handleViewDetails(annotation.id)}>
-                    <td className="checkbox-col">
-                      <input type="checkbox" onClick={(e) => e.stopPropagation()} />
-                    </td>
-                    <td>{annotation.cameraId}</td>
-                    <td>{annotation.id}</td>
-                    <td>{formatConfidenceScore(annotation.confidenceScore)}</td>
-                    <td>{annotation.defectCount}</td>
-                    <td>
-                      <span 
-                        className="status-tag"
-                        style={{ 
-                          backgroundColor: getStatusStyles(annotation.status).backgroundColor,
-                          color: getStatusStyles(annotation.status).color
-                        }}
-                      >
-                        {getStatusStyles(annotation.status).text}
-                      </span>
-                    </td>
-                    <td>
-                      <button 
-                        className="delete-btn"
-                        onClick={(e) => handleDeleteClick(annotation.id, e)}
-                      >
-                        <span>Delete</span>
-                      </button>
-                    </td>
-                  </tr>
-                ))}
-                
-                {annotations.length === 0 && (
-                  <tr>
-                    <td colSpan="7" className="no-data-message">
-                      표시할 데이터가 없습니다.
-                    </td>
-                  </tr>
-                )}
-              </tbody>
-            </table>
-          </div>
+          <AnnotationTable 
+            annotations={annotations}
+            onViewDetails={handleViewDetails}
+            onDelete={handleDelete}
+          />
         </div>
       </div>
     </div>
